PWN-8900 - comment fixes
