+++ app/src/main/java/org/p2p/wallet/striga/wallet/interactor/StrigaWalletInteractor.kt
@@ -10,9 +10,13 @@ import org.p2p.wallet.striga.wallet.models.StrigaWhitelistedAddressItem
 import org.p2p.wallet.striga.wallet.models.ids.StrigaAccountId
 import org.p2p.wallet.striga.wallet.models.ids.StrigaWhitelistedAddressId
 import org.p2p.wallet.striga.wallet.repository.StrigaWalletRepository
+import org.p2p.wallet.striga.wallet.repository.StrigaWhitelistAddressesRepository
+import org.p2p.wallet.striga.wallet.repository.StrigaWithdrawalsRepository
 
 class StrigaWalletInteractor(
-    private val repository: StrigaWalletRepository,
+    private val walletRepository: StrigaWalletRepository,
+    private val withdrawalsRepository: StrigaWithdrawalsRepository,
+    private val whitelistAddressesRepository: StrigaWhitelistAddressesRepository
 ) {
 
     private class StrigaEuroAccountNotFound : Throwable()
@@ -23,11 +27,11 @@ class StrigaWalletInteractor(
 
     @Throws(Throwable::class)
     suspend fun getFiatAccountDetails(): StrigaFiatAccountDetails {
-        val eurAccountId = repository.getUserWallet()
+        val eurAccountId = walletRepository.getUserWallet()
             .map { it.eurAccount?.accountId }
             .unwrap()
             ?: throw StrigaEuroAccountNotFound()
-        return repository.getFiatAccountDetails(eurAccountId).unwrap()
+        return walletRepository.getFiatAccountDetails(eurAccountId).unwrap()
     }
 
     /**
@@ -39,7 +43,7 @@ class StrigaWalletInteractor(
         amount: BigInteger,
     ): StrigaDataLayerResult<StrigaInitWithdrawalDetails> {
 
-        return repository.initiateOnchainWithdrawal(
+        return withdrawalsRepository.initiateOnchainWithdrawal(
             sourceAccountId = sourceAccountId,
             whitelistedAddressId = whitelistedAddressId,
             amountInUnits = amount,
@@ -51,14 +55,14 @@ class StrigaWalletInteractor(
         currency: StrigaNetworkCurrency,
         label: String?
     ): StrigaDataLayerResult<StrigaWhitelistedAddressItem> {
-        return repository.whitelistAddress(
+        return whitelistAddressesRepository.whitelistAddress(
             address = address,
             currency = currency,
-            label = label,
+            addressLabel = label,
         )
     }
 
     suspend fun getWhitelistedAddresses(): StrigaDataLayerResult<List<StrigaWhitelistedAddressItem>> {
-        return repository.getWhitelistedAddresses()
+        return whitelistAddressesRepository.getWhitelistedAddresses()
     }
 }
