+++ app/src/main/java/org/p2p/wallet/striga/wallet/repository/impl/StrigaWalletInMemoryRepository.kt
@@ -1,14 +1,14 @@
-package org.p2p.wallet.striga.wallet.repository
+package org.p2p.wallet.striga.wallet.repository.impl
 
 import org.p2p.wallet.striga.wallet.models.StrigaFiatAccountDetails
 import org.p2p.wallet.striga.wallet.models.StrigaUserWallet
 
 class StrigaWalletInMemoryRepository {
     var fiatAccountDetails: StrigaFiatAccountDetails? = null
-    var userStrigaWallet: StrigaUserWallet? = null
+    var userWallet: StrigaUserWallet? = null
 
     fun clear() {
         fiatAccountDetails = null
-        userStrigaWallet = null
+        userWallet = null
     }
 }
