+++ app/src/main/java/org/p2p/wallet/striga/wallet/repository/mapper/StrigaWalletMapper.kt
@@ -1,8 +1,11 @@
-package org.p2p.wallet.striga.wallet.repository
+package org.p2p.wallet.striga.wallet.repository.mapper
 
+import org.p2p.wallet.striga.wallet.api.response.StrigaEnrichFiatAccountResponse
 import org.p2p.wallet.striga.wallet.api.response.StrigaUserWalletAccountResponse
 import org.p2p.wallet.striga.wallet.api.response.StrigaUserWalletDetailsResponse
 import org.p2p.wallet.striga.wallet.api.response.StrigaUserWalletsResponse
+import org.p2p.wallet.striga.wallet.models.StrigaFiatAccountDetails
+import org.p2p.wallet.striga.wallet.models.StrigaFiatAccountStatus
 import org.p2p.wallet.striga.wallet.models.StrigaUserWallet
 import org.p2p.wallet.striga.wallet.models.StrigaUserWalletAccount
 import org.p2p.wallet.striga.wallet.models.StrigaWalletAccountBankLink
@@ -10,23 +13,44 @@ import org.p2p.wallet.striga.wallet.models.StrigaWalletAccountCurrency
 import org.p2p.wallet.striga.wallet.models.ids.StrigaAccountId
 import org.p2p.wallet.striga.wallet.models.ids.StrigaWalletId
 
-class StrigaUserWalletsMapper {
+class StrigaWalletMapper {
+
     private companion object {
         private const val EUR_ACCOUNT_NAME = "EUR"
         private const val USDC_ACCOUNT_NAME = "USDC"
         private const val UNLINKED_BANK_ACCOUNT_VALUE = "UNLINKED"
     }
 
+    fun fromNetwork(response: StrigaEnrichFiatAccountResponse): StrigaFiatAccountDetails = with(response) {
+        StrigaFiatAccountDetails(
+            currency = currency,
+            status = StrigaFiatAccountStatus.from(status),
+            internalAccountId = internalAccountId,
+            bankName = bankName,
+            bankCountry = bankCountry,
+            bankAddress = bankAddress,
+            bankAccountHolderName = bankAccountHolderName,
+            iban = iban,
+            bic = bic,
+            accountNumber = accountNumber,
+            provider = provider,
+            paymentType = paymentType,
+            isDomesticAccount = isDomesticAccount,
+            routingCodeEntries = routingCodeEntries,
+            payInReference = payInReference,
+        )
+    }
+
     fun fromNetwork(userId: String, response: StrigaUserWalletsResponse): StrigaUserWallet {
         require(response.wallets.isNotEmpty()) {
             "Wallets should be not empty: they are created when user $userId is created"
         }
+        // no support for multiple wallets so we get first
         val activeWallet: StrigaUserWalletDetailsResponse = response.wallets.first()
 
         return StrigaUserWallet(
             userId = userId,
             walletId = StrigaWalletId(activeWallet.walletId),
-            // no support for multiple wallets so we get first
             accounts = activeWallet.accountCurrencyToDetails.map(::toDomain)
         )
     }
